Move gate_sequence_product to qutip_qip.circuit
Add deprecation warning.
