+++ src/qutip_qip/qasm.py
@@ -8,13 +8,12 @@ import warnings
 from math import pi
 import numpy as np
 
-from .circuit import QubitCircuit
+from .circuit import QubitCircuit, gate_sequence_product
 from .operations import (
     controlled_gate,
     qasmu_gate,
     rz,
     snot,
-    gate_sequence_product,
 )
