+++ tests/test_device.py
@@ -7,8 +7,8 @@ import numpy as np
 import pytest
 
 import qutip
-from qutip_qip.circuit import QubitCircuit
-from qutip_qip.operations import Gate, gate_sequence_product
+from qutip_qip.circuit import QubitCircuit, gate_sequence_product
+from qutip_qip.operations import Gate
 from qutip_qip.device import (DispersiveCavityQED, LinearSpinChain,
                                 CircularSpinChain, SCQubits)
