+++ tests/test_qft.py
@@ -1,6 +1,6 @@
 from numpy.testing import assert_, assert_equal, assert_string_equal, run_module_suite
 from qutip_qip.algorithms.qft import qft, qft_steps, qft_gate_sequence
-from qutip_qip.operations import gate_sequence_product
+from qutip_qip.circuit import gate_sequence_product
 
 
 class TestQFT:
