+++ tests/test_scheduler.py
@@ -1,9 +1,9 @@
 import pytest
 from copy import deepcopy
 
-from qutip_qip.circuit import QubitCircuit
+from qutip_qip.circuit import QubitCircuit, gate_sequence_product
 from qutip_qip.compiler import Instruction, Scheduler
-from qutip_qip.operations import Gate, gate_sequence_product
+from qutip_qip.operations import Gate
 from qutip import process_fidelity, qeye, tracedist
